[#113576027] PHP Style Guide part 1 - Coding Standards
This covers the same ground as PSR-1 and is roughly the same except for a couple of areas where we can be more emphatic and opinionated. Those areas relate to Autoloading and formatting for properties. Both of those areas probably need some discussion before we finalize things for this section.